added unit tests around getSeleniumHub
