+++ lib/check-started.js
@@ -1,7 +1,7 @@
 module.exports = checkStarted;
 
 var request = require('request').defaults({json: true});
-var URI = require('URIjs');
+var getSeleniumHub = require('./get-selenium-hub.js');
 
 function checkStarted(seleniumArgs, cb) {
   var retries = 0;
@@ -31,23 +31,4 @@ function checkStarted(seleniumArgs, cb) {
   }
 
   hasStarted();
-}
-
-function getSeleniumHub(seleniumArgs) {
-  var port = 4444;
-  var hostname = 'localhost';
-  var portArg = seleniumArgs.indexOf('-port');
-  var hubArg = seleniumArgs.indexOf('-hub');
-
-  if (hubArg !== -1) {
-    hubURI = new URI(seleniumArgs[hubArg + 1]);
-    hostname = hubURI.hostname();
-    port = hubURI.port();
-  }
-
-  if (portArg !== -1) {
-    port = seleniumArgs[portArg + 1];
-  }
-
-  return 'http://' + hostname + ':' + port + '/wd/hub/status';
 }
